refactor: relocate dedicated components of `Works`
